+++ src/libs/antares/study/scenario-builder/TSnumberData.h
@@ -154,23 +154,6 @@ inline CString<512, false> solarTSNumberData::get_prefix() const
     return "s,";
 }
 
-// =====================
-// Hydro ...
-// =====================
-
-class hydroTSNumberData : public TSNumberData
-{
-public:
-    bool apply(Study& study) override;
-    CString<512, false> get_prefix() const override;
-    uint get_tsGenCount(const Study& study) const override;
-};
-
-inline CString<512, false> hydroTSNumberData::get_prefix() const
-{
-    return "h,";
-}
-
 } // namespace Antares::Data::ScenarioBuilder
 
 #endif // __LIBS_STUDY_SCENARIO_BUILDER_DATA_TS_NUMBER_H__
